crispy_forms added

+++ BLOG_APP_PROJECT_1/src/blog/views.py
@@ -1,5 +1,6 @@
 import imp
 from multiprocessing import context
+from django.http import HttpResponse
 import re
 from django.shortcuts import render, redirect
 from .models import Post, Like
@@ -53,8 +54,8 @@ def post_detail(request, slug):
 def post_update(request, slug):
     obj = get_object_or_404(Post, slug=slug)
     form = PostForm(request.POST or None, request.FILES or None, instance=obj)
-#     if request.user.id != obj.author.id:
-#         # return HttpResponse("You are not authorized!")
+    if request.user.id != obj.author.id:
+        return HttpResponse("You are not authorized!")
 #         messages.warning(request, "You are not a writer of this post !")
 #         return redirect("blog:list")
     if form.is_valid():
@@ -70,10 +71,10 @@ def post_update(request, slug):
 # @login_required()
 def post_delete(request, slug):
     obj = get_object_or_404(Post, slug=slug)
-    # if request.user.id != obj.author.id:
-#         # return HttpResponse("You are not authorized!")
-#         messages.warning(request, "You are not a writer of this post !")
-#         return redirect("blog:list")
+    if request.user.id != obj.author.id: #güvenliği sağlamak için:başkası delete yapmaması için
+        return HttpResponse("You are not authorized!")
+        # messages.warning(request, "You are not a writer of this post !")
+        # return redirect("blog:list")
     if request.method == "POST":
         obj.delete()
 #         messages.success(request, "Post deleted !!")
